+++ source/tricky-pipe/src/error.rs
