+++ source/tricky-pipe/src/static_impl.rs
